backend/pyproject.toml
[tool.poetry]
name = "Chatroom"
version = "0.1.0"
description = ""
authors = ["Mateusz Ostalecki"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.4"
channels = "^3.0.2"
Pillow = "^8.0.1"
psycopg2 = "^2.8.6"
channels_redis = "^3.2.0"
ipython = "^7.19.0"
djangorestframework-simplejwt = "^4.6.0"
djangorestframework = "^3.12.2"
djangorestframework-camel-case = "^1.2.0"
celery = "^5.0.5"
redis = "^3.5.3"
drf-yasg = "^1.20.0"
gunicorn = "^20.0.4"
python-decouple = "^3.4"
django-cors-headers = "^3.7.0"
django-ipware = "^3.0.2"
pytest-asyncio = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
factory_boy = "^3.2.0"
freezegun = "^1.1.0"
pytest-django = "^4.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
